Extract nested class
